Reject tokens with unrecognized claims
JWT lacks a mechanism for specifying critical claims as JWS has
with header 'crit' field; JWT spec suggests to ignore unrecognized
claims and that is what a generic JWT library should do.

However, conjur-rack is a specific application; for security reasons
it's best to reject tokens which we don't totally understand, lest the
requester is allowed to do what she is not in fact authorized to.

If need be, this can be extended in the future with 'crit'-like
mechanism for specifying optional claims.

Related: https://github.com/cyberark/slosilo/issues/12